protect private, modify detail view

+++ controllers/videoController.js
@@ -59,8 +59,10 @@ export const getEditVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    console.log(id);
     const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
     res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     console.log(error);
@@ -85,6 +87,10 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
     await Video.findOneAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
